+++ src/fifo_v2.sv
@@ -16,7 +16,9 @@ module fifo_v2 #(
     parameter int unsigned DEPTH        = 8,    // depth can be arbitrary from 0 to 2**32
     parameter int unsigned ALM_EMPTY_TH = 1,    // almost empty threshold (when to assert alm_empty_o)
     parameter int unsigned ALM_FULL_TH  = 1,    // almost full threshold (when to assert alm_full_o)
-    parameter type dtype                = logic [DATA_WIDTH-1:0]
+    parameter type dtype                = logic [DATA_WIDTH-1:0],
+    // DO NOT OVERWRITE THIS PARAMETER
+    parameter int unsigned ADDR_DEPTH   = (DEPTH > 1) ? $clog2(DEPTH) : 1
 )(
     input  logic  clk_i,            // Clock
     input  logic  rst_ni,           // Asynchronous reset active low
@@ -34,124 +36,39 @@ module fifo_v2 #(
     output dtype  data_o,           // output data
     input  logic  pop_i             // pop head from queue
 );
-    // local parameter
-    // FIFO depth - handle the case of pass-through, synthesizer will do constant propagation
-    localparam int unsigned FIFO_DEPTH = (DEPTH > 0) ? DEPTH : 1;
-    localparam int unsigned ADDR_DEPTH = (DEPTH > 1) ? $clog2(DEPTH) : 1;
-    // clock gating control
-    logic gate_clock;
-    // pointer to the read and write section of the queue
-    logic [ADDR_DEPTH - 1:0] read_pointer_n, read_pointer_q, write_pointer_n, write_pointer_q;
-    // keep a counter to keep track of the current queue status
-    logic [ADDR_DEPTH:0] status_cnt_n, status_cnt_q; // this integer will be truncated by the synthesis tool
-    // actual memory
-    dtype [FIFO_DEPTH - 1:0] mem_n, mem_q;
 
+    logic [ADDR_DEPTH-1:0] usage;
+
+    // generate threshold parameters
     if (DEPTH == 0) begin
-        assign empty_o     = ~push_i;
-        assign full_o      = ~pop_i;
         assign alm_full_o  = 1'b0; // that signal does not make any sense in a FIFO of depth 0
         assign alm_empty_o = 1'b0; // that signal does not make any sense in a FIFO of depth 0
     end else begin
-        assign full_o       = (status_cnt_q == FIFO_DEPTH);
-        assign empty_o      = (status_cnt_q == 0) & ~(FALL_THROUGH & push_i);
-        assign alm_full_o   = (status_cnt_q >= ALM_FULL_TH);
-        assign alm_empty_o  = (status_cnt_q <= ALM_EMPTY_TH);
-    end
-    // status flags
-
-    // read and write queue logic
-    always_comb begin : read_write_comb
-        // default assignment
-        read_pointer_n  = read_pointer_q;
-        write_pointer_n = write_pointer_q;
-        status_cnt_n    = status_cnt_q;
-        data_o          = (DEPTH == 0) ? data_i : mem_q[read_pointer_q];
-        mem_n           = mem_q;
-        gate_clock      = 1'b1;
-
-        // push a new element to the queue
-        if (push_i && ~full_o) begin
-            // push the data onto the queue
-            mem_n[write_pointer_q] = data_i;
-            // un-gate the clock, we want to write something
-            gate_clock = 1'b0;
-            // increment the write counter
-            if (write_pointer_q == FIFO_DEPTH-1)
-                write_pointer_n = '0;
-            else
-                write_pointer_n = write_pointer_q + 1;
-            // increment the overall counter
-            status_cnt_n    = status_cnt_q + 1;
-        end
-
-        if (pop_i && ~empty_o) begin
-            // read from the queue is a default assignment
-            // but increment the read pointer...
-            if (read_pointer_n == FIFO_DEPTH-1)
-                read_pointer_n = '0;
-            else
-                read_pointer_n = read_pointer_q + 1;
-            // ... and decrement the overall count
-            status_cnt_n   = status_cnt_q - 1;
-        end
-
-        // keep the count pointer stable if we push and pop at the same time
-        if (push_i && pop_i &&  ~full_o && ~empty_o)
-            status_cnt_n   = status_cnt_q;
-
-        // FIFO is in pass through mode -> do not change the pointers
-        if (FALL_THROUGH && (status_cnt_q == 0) && push_i && pop_i) begin
-            data_o = data_i;
-            status_cnt_n = status_cnt_q;
-            read_pointer_n = read_pointer_q;
-            write_pointer_n = write_pointer_q;
-        end
+        assign alm_full_o   = (usage >= ALM_FULL_TH[ADDR_DEPTH-1:0]);
+        assign alm_empty_o  = (usage <= ALM_EMPTY_TH[ADDR_DEPTH-1:0]);
     end
 
-    // sequential process
-    always_ff @(posedge clk_i or negedge rst_ni) begin
-        if(~rst_ni) begin
-            read_pointer_q  <= '0;
-            write_pointer_q <= '0;
-            status_cnt_q    <= '0;
-        end else begin
-            if (flush_i) begin
-                read_pointer_q  <= '0;
-                write_pointer_q <= '0;
-                status_cnt_q    <= '0;
-             end else begin
-                read_pointer_q  <= read_pointer_n;
-                write_pointer_q <= write_pointer_n;
-                status_cnt_q    <= status_cnt_n;
-            end
-        end
-    end
-
-    always_ff @(posedge clk_i or negedge rst_ni) begin
-        if(~rst_ni) begin
-            mem_q <= '0;
-        end else if (!gate_clock) begin
-            mem_q <= mem_n;
-        end
-    end
-
-//pragma translate_off
-`ifndef VERILATOR
-    initial begin
-        assert (ALM_FULL_TH <= DEPTH)  else $error("ALM_FULL_TH can't be larger than the DEPTH.");
-        assert (ALM_EMPTY_TH <= DEPTH) else $error("ALM_EMPTY_TH can't be larger than the DEPTH.");
-        assert (DEPTH > 0)             else $error("DEPTH must be greater than 0.");
-    end
-
-    full_write : assert property(
-        @(posedge clk_i) disable iff (~rst_ni) (full_o |-> ~push_i))
-        else $fatal (1,"Trying to push new data although the FIFO is full.");
-
-    empty_read : assert property(
-        @(posedge clk_i) disable iff (~rst_ni) (empty_o |-> ~pop_i))
-        else $fatal (1,"Trying to pop data although the FIFO is empty.");
-`endif
-//pragma translate_on
+    fifo_v3 #(
+        .FALL_THROUGH,
+        .DATA_WIDTH,
+        .DEPTH,
+        .ALM_EMPTY_TH,
+        .ALM_FULL_TH,
+        .dtype ( dtype )
+    ) i_fifo_v3 (
+        .clk_i,
+        .rst_ni,
+        .flush_i,
+        .testmode_i,
+        .full_o,
+        .empty_o,
+        .usage_o (usage),
+        .alm_full_o,
+        .alm_empty_o,
+        .data_i,
+        .push_i,
+        .data_o,
+        .pop_i
+    );
 
 endmodule // fifo_v2